requirements.txt
Faker==8.1.0
django_rest_swagger==2.2.0
validate_email==1.3
drf_yasg==1.20.0
Django==3.2
djangorestframework_simplejwt==4.6.0
djangorestframework==3.12.4
PyJWT==2.0.1
python-decouple==3.4
psycopg2-binary>=2.8
django-crontab==0.7.1
facebook-sdk==3.1.0
